bchain: refactor ledger so that it has 'apis' layer
Functionalities are to be defined under 'apis' from now on.
