+++ source/saksaha/src/blockchain/ledger/apis/block.rs
@@ -1,0 +1,265 @@
+use crate::blockchain::{
+    blockchain::TxValue,
+    ledger::{block_columns, Hashable},
+    BlockValue, Hash,
+};
+use database::KeyValueDatabase;
+use logger::tinfo;
+use rocksdb::{
+    DBRawIteratorWithThreadMode, DBWithThreadMode, SingleThreaded, WriteBatch,
+};
+use sha3::{Digest, Sha3_256};
+
+#[inline]
+pub(crate) async fn get_block(
+    // &self,
+    ledger_db: &KeyValueDatabase,
+    block_hash: &Hash,
+) -> Result<BlockValue, String> {
+    // let db = &self.ledger_db.db;
+    let db = &ledger_db.db;
+
+    let cf_handle = match db.cf_handle(block_columns::CREATED_AT) {
+        Some(h) => h,
+        None => {
+            return Err(format!(
+                "Fail to open ledger columns {}",
+                block_columns::CREATED_AT
+            ));
+        }
+    };
+
+    let created_at = match db.get_cf(cf_handle, &block_hash.hash) {
+        Ok(val) => match val {
+            Some(v) => match std::str::from_utf8(&v) {
+                Ok(vs) => vs.to_string(),
+                Err(err) => {
+                    return Err(format!("Invalid utf8 given, err: {}", err,));
+                }
+            },
+            None => {
+                return Err(format!(
+                    "No matched value with tx_hash in {}, {}",
+                    block_columns::CREATED_AT,
+                    &block_hash.hash,
+                ));
+            }
+        },
+        Err(err) => {
+            return Err(format!(
+                "Fail to get value from ledger columns, column: {}, \
+                    err: {}",
+                block_columns::CREATED_AT,
+                err,
+            ));
+        }
+    };
+
+    let cf_handle = match db.cf_handle(block_columns::TX_POOL) {
+        Some(h) => h,
+        None => {
+            return Err(format!(
+                "Fail to open ledger columns {}",
+                block_columns::TX_POOL
+            ));
+        }
+    };
+    let get_cf_handle = db.get_cf(cf_handle, &block_hash.hash);
+
+    let tx_pool = match get_cf_handle.as_ref() {
+        Ok(val) => match val.as_ref() {
+            Some(v) => {
+                let th: Vec<String> = serde_json::from_slice(&v).unwrap();
+                th
+            }
+            None => {
+                return Err(format!(
+                    "No matched value with tx_hash in {}, {}",
+                    block_columns::TX_POOL,
+                    &block_hash.hash,
+                ));
+            }
+        },
+        Err(err) => {
+            return Err(format!(
+                "Fail to get value from ledger columns, column: {}, \
+                    err: {}",
+                block_columns::TX_POOL,
+                err,
+            ));
+        }
+    };
+
+    let cf_handle = match db.cf_handle(block_columns::SIG_VEC) {
+        Some(h) => h,
+        None => {
+            return Err(format!(
+                "Fail to open ledger columns {}",
+                block_columns::SIG_VEC
+            ));
+        }
+    };
+    let get_cf_handle = db.get_cf(cf_handle, &block_hash.hash);
+
+    let sig_vec = match get_cf_handle.as_ref() {
+        Ok(val) => match val.as_ref() {
+            Some(v) => {
+                let th: Vec<String> = serde_json::from_slice(&v).unwrap();
+                th
+            }
+            None => {
+                return Err(format!(
+                    "No matched value with tx_hash in {}, {}",
+                    block_columns::SIG_VEC,
+                    &block_hash.hash,
+                ));
+            }
+        },
+        Err(err) => {
+            return Err(format!(
+                "Fail to get value from ledger columns, column: {}, \
+                    err: {}",
+                block_columns::SIG_VEC,
+                err,
+            ));
+        }
+    };
+
+    let cf_handle = match db.cf_handle(block_columns::HEIGHT) {
+        Some(h) => h,
+        None => {
+            return Err(format!(
+                "Fail to open ledger columns {}",
+                block_columns::HEIGHT
+            ));
+        }
+    };
+
+    let height = match db.get_cf(cf_handle, &block_hash.hash) {
+        Ok(val) => match val {
+            Some(v) => match std::str::from_utf8(&v) {
+                Ok(vs) => vs.to_string(),
+                Err(err) => {
+                    return Err(format!("Invalid utf8 given, err: {}", err,));
+                }
+            },
+            None => {
+                return Err(format!(
+                    "No matched value with tx_hash in {}, {}",
+                    block_columns::HEIGHT,
+                    &block_hash.hash,
+                ));
+            }
+        },
+        Err(err) => {
+            return Err(format!(
+                "Fail to get value from ledger columns, column: {}, \
+                    err: {}",
+                block_columns::HEIGHT,
+                err,
+            ));
+        }
+    };
+
+    let bv = BlockValue {
+        tx_pool,
+        sig_vec,
+        created_at,
+        height,
+    };
+    println!("requested blockvalue: {:?}", &bv);
+    Ok(bv)
+}
+
+#[inline]
+pub(crate) async fn write_block(
+    // &self,
+    ledger_db: &KeyValueDatabase,
+    block_value: BlockValue,
+) -> Result<Hash, String> {
+    // let db = &self.ledger_db.db;
+    let db = &ledger_db.db;
+
+    let mut batch = WriteBatch::default();
+
+    let block_hash = match block_value.get_hash() {
+        Ok(hash) => hash,
+        Err(_) => return Err(format!("Failed to get hash from block_value")),
+    };
+
+    println!(
+        "write_block(): created_at: {}, block_hash: {:?}",
+        block_value.created_at, block_hash
+    );
+
+    let cf_handle = match db.cf_handle(block_columns::CREATED_AT) {
+        Some(h) => h,
+        None => {
+            return Err(format!(
+                "Fail to open ledger columns {}",
+                block_columns::CREATED_AT
+            ))
+        }
+    };
+    batch.put_cf(cf_handle, &block_hash.hash, block_value.created_at);
+
+    let cf_handle = match db.cf_handle(block_columns::SIG_VEC) {
+        Some(h) => h,
+        None => {
+            return Err(format!(
+                "Fail to open ledger columns {}",
+                block_columns::SIG_VEC
+            ))
+        }
+    };
+    let ser_sig_vec = match serde_json::to_string(&block_value.sig_vec) {
+        Ok(v) => v,
+        Err(err) => {
+            return Err(format!(
+                "Cannot serialize {}, err: {}",
+                block_columns::SIG_VEC,
+                err
+            ))
+        }
+    };
+    batch.put_cf(cf_handle, &block_hash.hash, ser_sig_vec);
+
+    let cf_handle = match db.cf_handle(block_columns::HEIGHT) {
+        Some(h) => h,
+        None => {
+            return Err(format!(
+                "Fail to open ledger columns {}",
+                block_columns::HEIGHT
+            ))
+        }
+    };
+    batch.put_cf(cf_handle, &block_hash.hash, block_value.height);
+
+    let cf_handle = match db.cf_handle(block_columns::TX_POOL) {
+        Some(h) => h,
+        None => {
+            return Err(format!(
+                "Fail to open ledger columns {}",
+                block_columns::TX_POOL
+            ))
+        }
+    };
+    let ser_tx_pool = match serde_json::to_string(&block_value.tx_pool) {
+        Ok(v) => v,
+        Err(err) => {
+            return Err(format!(
+                "Cannot serialize {}, err: {}",
+                block_columns::TX_POOL,
+                err
+            ))
+        }
+    };
+    batch.put_cf(cf_handle, &block_hash.hash, ser_tx_pool);
+
+    match db.write(batch) {
+        Ok(_) => return Ok(block_hash),
+        Err(err) => {
+            return Err(format!("Fail to write on ledger db, err: {}", err))
+        }
+    }
+}
