+++ source/saksaha/src/blockchain/ledger/ledger.rs
@@ -1,11 +1,9 @@
-use super::{db, tx_columns};
+use super::{apis, db, tx_columns};
 use crate::blockchain::{blockchain::TxValue, BlockValue, Hash};
 use database::KeyValueDatabase;
 use db::block_columns;
 use logger::tinfo;
-use rocksdb::{
-    DBRawIteratorWithThreadMode, DBWithThreadMode, SingleThreaded, WriteBatch,
-};
+use rocksdb::{DBRawIteratorWithThreadMode, DBWithThreadMode, SingleThreaded};
 use sha3::{Digest, Sha3_256};
 
 pub(crate) struct Ledger {
@@ -35,389 +33,28 @@ impl Ledger {
         &self,
         tx_value: TxValue,
     ) -> Result<Hash, String> {
-        let db = &self.ledger_db.db;
-
-        let mut batch = WriteBatch::default();
-
-        let tx_hash = match tx_value.get_hash() {
-            Ok(hash) => hash,
-            Err(_) => {
-                return Err(format!("Failed to get hash from tx_value"))
-            }
-        };
-
-        let cf_handle = match db.cf_handle(tx_columns::CREATED_AT) {
-            Some(h) => h,
-            None => {
-                return Err(format!(
-                    "Fail to open ledger columns {}",
-                    tx_columns::CREATED_AT
-                ))
-            }
-        };
-        batch.put_cf(cf_handle, &tx_hash.hash, tx_value.created_at);
-
-        let cf_handle = match db.cf_handle(tx_columns::DATA) {
-            Some(h) => h,
-            None => {
-                return Err(format!(
-                    "Fail to open ledger columns {}",
-                    tx_columns::DATA
-                ))
-            }
-        };
-        batch.put_cf(cf_handle, &tx_hash.hash, tx_value.data);
-
-        let cf_handle = match db.cf_handle(tx_columns::PI) {
-            Some(h) => h,
-            None => {
-                return Err(format!(
-                    "Fail to open ledger columns {}",
-                    tx_columns::PI
-                ))
-            }
-        };
-        batch.put_cf(cf_handle, &tx_hash.hash, tx_value.pi);
-
-        let cf_handle = match db.cf_handle(tx_columns::SIG_VEC) {
-            Some(h) => h,
-            None => {
-                return Err(format!(
-                    "Fail to open ledger columns {}",
-                    tx_columns::SIG_VEC
-                ))
-            }
-        };
-        batch.put_cf(cf_handle, &tx_hash.hash, tx_value.sig_vec);
-
-        match db.write(batch) {
-            Ok(_) => return Ok(tx_hash),
-            Err(err) => {
-                return Err(format!("Fail to write on ledger db, err: {}", err))
-            }
-        }
+        apis::write_tx(&self.ledger_db, tx_value).await
     }
 
     pub(crate) async fn read_tx(
         &self,
         tx_hash: &Hash,
     ) -> Result<TxValue, String> {
-        let db = &self.ledger_db.db;
-
-        let mut tx_value_result = vec![
-            String::from(""),
-            String::from(""),
-            String::from(""),
-            String::from(""),
-        ];
-
-        let tx_values_col = vec![
-            tx_columns::CREATED_AT,
-            tx_columns::DATA,
-            tx_columns::SIG_VEC,
-            tx_columns::PI,
-        ];
-
-        let tx_values_it_map = tx_values_col.iter().map(|cf_name| cf_name);
-
-        for (idx, cfn) in tx_values_it_map.enumerate() {
-            let cf_handle = match db.cf_handle(cfn) {
-                Some(h) => h,
-                None => {
-                    return Err(format!("Fail to open ledger columns {}", cfn));
-                }
-            };
-
-            tx_value_result[idx] = match db.get_cf(cf_handle, &tx_hash.hash) {
-                Ok(val) => match val {
-                    Some(v) => match std::str::from_utf8(&v) {
-                        Ok(vs) => vs.to_string(),
-                        Err(err) => {
-                            return Err(format!(
-                                "Invalid utf8 given, err: {}",
-                                err,
-                            ));
-                        }
-                    },
-                    None => {
-                        return Err(format!(
-                            "No matched value with tx_hash in {}, {}",
-                            cfn, &tx_hash.hash,
-                        ));
-                    }
-                },
-                Err(err) => {
-                    return Err(format!(
-                        "Fail to get value from ledger columns, column: {}, \
-                            err: {}",
-                        cfn, err,
-                    ));
-                }
-            };
-        }
-
-        Ok(TxValue {
-            created_at: tx_value_result[0].clone(),
-            data: tx_value_result[1].clone(),
-            sig_vec: tx_value_result[2].clone(),
-            pi: tx_value_result[3].clone(),
-        })
+        apis::read_tx(&self.ledger_db, tx_hash).await
     }
 
     pub(crate) async fn get_block(
         &self,
         block_hash: &Hash,
     ) -> Result<BlockValue, String> {
-        let db = &self.ledger_db.db;
-
-        let cf_handle = match db.cf_handle(block_columns::CREATED_AT) {
-            Some(h) => h,
-            None => {
-                return Err(format!(
-                    "Fail to open ledger columns {}",
-                    block_columns::CREATED_AT
-                ));
-            }
-        };
-
-        let created_at = match db.get_cf(cf_handle, &block_hash.hash) {
-            Ok(val) => match val {
-                Some(v) => match std::str::from_utf8(&v) {
-                    Ok(vs) => vs.to_string(),
-                    Err(err) => {
-                        return Err(format!(
-                            "Invalid utf8 given, err: {}",
-                            err,
-                        ));
-                    }
-                },
-                None => {
-                    return Err(format!(
-                        "No matched value with tx_hash in {}, {}",
-                        block_columns::CREATED_AT,
-                        &block_hash.hash,
-                    ));
-                }
-            },
-            Err(err) => {
-                return Err(format!(
-                    "Fail to get value from ledger columns, column: {}, \
-                        err: {}",
-                    block_columns::CREATED_AT,
-                    err,
-                ));
-            }
-        };
-
-        let cf_handle = match db.cf_handle(block_columns::TX_POOL) {
-            Some(h) => h,
-            None => {
-                return Err(format!(
-                    "Fail to open ledger columns {}",
-                    block_columns::TX_POOL
-                ));
-            }
-        };
-        let get_cf_handle = db.get_cf(cf_handle, &block_hash.hash);
-
-        let tx_pool = match get_cf_handle.as_ref() {
-            Ok(val) => match val.as_ref() {
-                Some(v) => {
-                    let th: Vec<String> = serde_json::from_slice(&v).unwrap();
-                    th
-                }
-                None => {
-                    return Err(format!(
-                        "No matched value with tx_hash in {}, {}",
-                        block_columns::TX_POOL,
-                        &block_hash.hash,
-                    ));
-                }
-            },
-            Err(err) => {
-                return Err(format!(
-                    "Fail to get value from ledger columns, column: {}, \
-                        err: {}",
-                    block_columns::TX_POOL,
-                    err,
-                ));
-            }
-        };
-
-        let cf_handle = match db.cf_handle(block_columns::SIG_VEC) {
-            Some(h) => h,
-            None => {
-                return Err(format!(
-                    "Fail to open ledger columns {}",
-                    block_columns::SIG_VEC
-                ));
-            }
-        };
-        let get_cf_handle = db.get_cf(cf_handle, &block_hash.hash);
-
-        let sig_vec = match get_cf_handle.as_ref() {
-            Ok(val) => match val.as_ref() {
-                Some(v) => {
-                    let th: Vec<String> = serde_json::from_slice(&v).unwrap();
-                    th
-                }
-                None => {
-                    return Err(format!(
-                        "No matched value with tx_hash in {}, {}",
-                        block_columns::SIG_VEC,
-                        &block_hash.hash,
-                    ));
-                }
-            },
-            Err(err) => {
-                return Err(format!(
-                    "Fail to get value from ledger columns, column: {}, \
-                        err: {}",
-                    block_columns::SIG_VEC,
-                    err,
-                ));
-            }
-        };
-
-        let cf_handle = match db.cf_handle(block_columns::HEIGHT) {
-            Some(h) => h,
-            None => {
-                return Err(format!(
-                    "Fail to open ledger columns {}",
-                    block_columns::HEIGHT
-                ));
-            }
-        };
-
-        let height = match db.get_cf(cf_handle, &block_hash.hash) {
-            Ok(val) => match val {
-                Some(v) => match std::str::from_utf8(&v) {
-                    Ok(vs) => vs.to_string(),
-                    Err(err) => {
-                        return Err(format!(
-                            "Invalid utf8 given, err: {}",
-                            err,
-                        ));
-                    }
-                },
-                None => {
-                    return Err(format!(
-                        "No matched value with tx_hash in {}, {}",
-                        block_columns::HEIGHT,
-                        &block_hash.hash,
-                    ));
-                }
-            },
-            Err(err) => {
-                return Err(format!(
-                    "Fail to get value from ledger columns, column: {}, \
-                        err: {}",
-                    block_columns::HEIGHT,
-                    err,
-                ));
-            }
-        };
-
-        let bv = BlockValue {
-            tx_pool,
-            sig_vec,
-            created_at,
-            height,
-        };
-        println!("requested blockvalue: {:?}", &bv);
-        Ok(bv)
+        apis::get_block(&self.ledger_db, block_hash).await
     }
 
     pub(crate) async fn write_block(
         &self,
         block_value: BlockValue,
     ) -> Result<Hash, String> {
-        let db = &self.ledger_db.db;
-
-        let mut batch = WriteBatch::default();
-
-        let block_hash = match block_value.get_hash() {
-            Ok(hash) => hash,
-            Err(_) => {
-                return Err(format!("Failed to get hash from block_value"))
-            }
-        };
-
-        println!(
-            "write_block(): created_at: {}, block_hash: {:?}",
-            block_value.created_at, block_hash
-        );
-
-        let cf_handle = match db.cf_handle(block_columns::CREATED_AT) {
-            Some(h) => h,
-            None => {
-                return Err(format!(
-                    "Fail to open ledger columns {}",
-                    block_columns::CREATED_AT
-                ))
-            }
-        };
-        batch.put_cf(cf_handle, &block_hash.hash, block_value.created_at);
-
-        let cf_handle = match db.cf_handle(block_columns::SIG_VEC) {
-            Some(h) => h,
-            None => {
-                return Err(format!(
-                    "Fail to open ledger columns {}",
-                    block_columns::SIG_VEC
-                ))
-            }
-        };
-        let ser_sig_vec = match serde_json::to_string(&block_value.sig_vec) {
-            Ok(v) => v,
-            Err(err) => {
-                return Err(format!(
-                    "Cannot serialize {}, err: {}",
-                    block_columns::SIG_VEC,
-                    err
-                ))
-            }
-        };
-        batch.put_cf(cf_handle, &block_hash.hash, ser_sig_vec);
-
-        let cf_handle = match db.cf_handle(block_columns::HEIGHT) {
-            Some(h) => h,
-            None => {
-                return Err(format!(
-                    "Fail to open ledger columns {}",
-                    block_columns::HEIGHT
-                ))
-            }
-        };
-        batch.put_cf(cf_handle, &block_hash.hash, block_value.height);
-
-        let cf_handle = match db.cf_handle(block_columns::TX_POOL) {
-            Some(h) => h,
-            None => {
-                return Err(format!(
-                    "Fail to open ledger columns {}",
-                    block_columns::TX_POOL
-                ))
-            }
-        };
-        let ser_tx_pool = match serde_json::to_string(&block_value.tx_pool) {
-            Ok(v) => v,
-            Err(err) => {
-                return Err(format!(
-                    "Cannot serialize {}, err: {}",
-                    block_columns::TX_POOL,
-                    err
-                ))
-            }
-        };
-        batch.put_cf(cf_handle, &block_hash.hash, ser_tx_pool);
-
-        match db.write(batch) {
-            Ok(_) => return Ok(block_hash),
-            Err(err) => {
-                return Err(format!("Fail to write on ledger db, err: {}", err))
-            }
-        }
+        apis::write_block(&self.ledger_db, block_value).await
     }
 
     pub fn iter(
@@ -431,11 +68,11 @@ impl Ledger {
     }
 }
 
-pub trait Hashing {
+pub trait Hashable {
     fn get_hash(&self) -> Result<Hash, String>;
 }
 
-impl Hashing for BlockValue {
+impl Hashable for BlockValue {
     fn get_hash(&self) -> Result<Hash, String> {
         let hash = {
             let mut h = Sha3_256::new();
@@ -458,7 +95,7 @@ impl Hashing for BlockValue {
     }
 }
 
-impl Hashing for TxValue {
+impl Hashable for TxValue {
     fn get_hash(&self) -> Result<Hash, String> {
         let hash = {
             let mut h = Sha3_256::new();
